[c] More 4.2 porting, SkeletonBinary, untested, probably broken

+++ spine-c/spine-c/include/spine/MeshAttachment.h
@@ -65,7 +65,7 @@ struct spMeshAttachment {
 
 	/* Nonessential. */
 	int edgesCount;
-	int *edges;
+	unsigned short *edges;
 	float width, height;
 };
